chainlit Q A/requirements.txt
langchain
chromadb
tiktoken
chainlit
openai
python-dotenv
